fix(d3d12): keep tile copy flag mask constexpr

+++ src/d3d12/d3d12_command_queue.cpp
@@ -9523,8 +9523,8 @@ private:
     }
 
     constexpr UINT kDirectionFlags =
-        D3D12_TILE_COPY_FLAG_LINEAR_BUFFER_TO_SWIZZLED_TILED_RESOURCE |
-        D3D12_TILE_COPY_FLAG_SWIZZLED_TILED_RESOURCE_TO_LINEAR_BUFFER;
+        static_cast<UINT>(D3D12_TILE_COPY_FLAG_LINEAR_BUFFER_TO_SWIZZLED_TILED_RESOURCE) |
+        static_cast<UINT>(D3D12_TILE_COPY_FLAG_SWIZZLED_TILED_RESOURCE_TO_LINEAR_BUFFER);
     if (static_cast<UINT>(record.flags) & ~kDirectionFlags) {
       WARN("D3D12CommandQueue: TODO CopyTiles unsupported flags flags=", record.flags);
       return;
